+++ changelog.md
@@ -1,3 +1,8 @@
+v0.2.0
+* Refactor app to use internal middlewares for handling operations
+* Refactor JSON decoding to a middleware
+* Remove timestamp option
+
 v0.1.3
 * Refactor config to use a struct
 * Update readme to cover HTTP pipelining